Rename some files.
